+++ packages/stream_video/test/src/call/call_test.dart
@@ -155,7 +155,7 @@ void main() {
         );
 
         internetStatusController.add(InternetStatus.disconnected);
-        await Future<void>.delayed(Duration(microseconds: 100));
+        await Future<void>.delayed(Duration.zero);
         internetStatusController.add(InternetStatus.connected);
         await Future<void>.delayed(Duration.zero);