+++ packages/stream_video_push_notification/lib/src/stream_video_push_params.g.dart
@@ -97,6 +97,7 @@ IOSParams _$IOSParamsFromJson(Map<String, dynamic> json) => IOSParams(
   supportsGrouping: json['supportsGrouping'] as bool?,
   supportsUngrouping: json['supportsUngrouping'] as bool?,
   ringtonePath: json['ringtonePath'] as String?,
+  includesCallsInRecents: json['includesCallsInRecents'] as bool?,
 );
 
 Map<String, dynamic> _$IOSParamsToJson(IOSParams instance) => <String, dynamic>{
@@ -117,6 +118,7 @@ Map<String, dynamic> _$IOSParamsToJson(IOSParams instance) => <String, dynamic>{
   'supportsGrouping': instance.supportsGrouping,
   'supportsUngrouping': instance.supportsUngrouping,
   'ringtonePath': instance.ringtonePath,
+  'includesCallsInRecents': instance.includesCallsInRecents,
 };
 
 MissedCallNotificationParams _$MissedCallNotificationParamsFromJson(