+++ packages/stream_video_push_notification/android/src/main/kotlin/io/getstream/video/flutter/stream_video_push_notification/StreamVideoPushNotificationPlugin.kt
@@ -375,7 +375,6 @@ override fun onDetachedFromActivityForConfigChanges() {
     }
 
     override fun onDetachedFromActivity() {
-        instance.context = null
         instance.activity = null
     }